Update: double precision for LU (Very Important)

+++ src/func/linsys.py
@@ -18,7 +18,7 @@ class completePartial(nn.Module):
         self.partial_ind = partial_ind
         self.other_ind = [i for i in range(num_var) if i not in partial_ind]
         # precompute LU decomposition
-        self.lu, self.pivots = lu_factor(A[:, self.other_ind])
+        self.lu, self.pivots = lu_factor(A[:, self.other_ind].double())
         # keys
         self.x_key = var_key
         self.b_key = rhs_key
@@ -30,10 +30,10 @@ class completePartial(nn.Module):
         """
         # get values
         x, b = data[self.x_key], data[self.b_key]
-        rhs = b - (x @ self.A[:, self.partial_ind].T)
+        rhs = (b - (x @ self.A[:, self.partial_ind].T)).double()
         # complete vars
         x_comp = torch.zeros(x.shape[0], self.num_var, device=x.device)
         x_comp[:, self.partial_ind] = x
-        x_comp[:, self.other_ind] = lu_solve(self.lu, self.pivots, rhs.T).T
+        x_comp[:, self.other_ind] = lu_solve(self.lu, self.pivots, rhs.T).T.float()
         data[self.out_key] = x_comp
         return data
